Fix memory leak
Before, the text of every button was being rerendered every frame, which
caused a memory leak. Now I have put all the text specific data in a
Text struct. The Text struct had a method that rerenders the text, so
you can still rerender the text when you find it necessary, but it
doesn't happen every frame.